Implemented PMTrackerDatabase public functions with hibernate
 * Adopted PMTrackerDatabase Tests

+++ data-collection/src/main/java/refactoringml/PMTrackerDatabase.java
@@ -5,37 +5,27 @@ import refactoringml.db.CommitMetaData;
 import refactoringml.db.Database;
 import refactoringml.db.HibernateConfig;
 import javax.persistence.*;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 public class PMTrackerDatabase {
-	//Map class files onto their original process metrics.
-	@ElementCollection
-	@CollectionTable(name = "PMTrackerDatabase")
-	@MapKeyColumn(name = "filePath")
-	@Column(name = "pmTracker")
-	private Map<String, ProcessMetricTracker> database;
-	//All commit thresholds for this project for considering a class file as stable
 	@Transient
 	private List<Integer> commitThresholds;
 
 	@Transient
-	private final String databasePath =	"jdbc:mysql://localhost/pmDatabase";
+	private final String databasePath =	"jdbc:mysql://localhost/refactoringtest";
 	@Transient
 	private final String databaseUsername = "root";
 	@Transient
 	private final String databasePassword = "root";
 	@Transient
-	private Database db;
+	public Database db;
 
 	public PMTrackerDatabase(List<Integer> commitThresholds) {
 		SessionFactory sf = new HibernateConfig().getSessionFactory(databasePath, databaseUsername, databasePassword);
 		db = new Database(sf);
 
 		this.commitThresholds = commitThresholds;
-		this.database = new HashMap<>();
 	}
 
 	//Empty the database and close the hibernate SessionFactory
@@ -47,12 +37,12 @@ public class PMTrackerDatabase {
 	//public interaction
 	//Retrieve the process metrics tracker for the given fileName
 	public ProcessMetricTracker find(String fileName) {
-		return database.get(fileName);
+		return db.find(fileName, ProcessMetricTracker.class);
 	}
 
 	//Find all stable instances in the database
 	public List<ProcessMetricTracker> findStableInstances() {
-		return database.values().stream()
+		return db.getAll("ProcessMetricTracker", ProcessMetricTracker.class).stream()
 				.filter(pmTracker -> pmTracker.calculateStability(commitThresholds))
 				.collect(Collectors.toList());
 	}
@@ -64,41 +54,63 @@ public class PMTrackerDatabase {
 	2. Rename Person.java to Human.java: Human -> People_ProcessMetrics
 	 */
 	public ProcessMetricTracker renameFile(String oldFileName, String newFileName, CommitMetaData commitMetaData){
-		ProcessMetricTracker pmTracker = new ProcessMetricTracker(database.getOrDefault(oldFileName, new ProcessMetricTracker(newFileName, commitMetaData)));
+		if(oldFileName.equals(newFileName))
+			return db.find(oldFileName, ProcessMetricTracker.class);
+
+		ProcessMetricTracker oldPMTracker = db.find(oldFileName, ProcessMetricTracker.class);
+		if(oldPMTracker == null)
+			return null;
+
+		ProcessMetricTracker pmTracker = new ProcessMetricTracker(oldPMTracker);
 		pmTracker.setFileName(newFileName);
 
-		ProcessMetricTracker oldPMTracker = removeFile(oldFileName);
+		removeFile(oldFileName);
+		db.persist(pmTracker);
 
-		database.put(newFileName, pmTracker);
 		return oldPMTracker;
 	}
 
 	//Remove the given fileName from the process metrics database
 	//Returns the old process metrics tracker of the deleted class file, if any existed in the database
 	public ProcessMetricTracker removeFile(String fileName){
-		return database.remove(fileName);
+		return db.remove(fileName, ProcessMetricTracker.class);
 	}
 
 	//Report a commit changing the given class file, the in memory database is updated accordingly
 	public void reportChanges(String fileName, CommitMetaData commitMetaData, String authorName, int linesAdded, int linesDeleted) {
-		ProcessMetricTracker pmTracker = database.getOrDefault(fileName, new ProcessMetricTracker(fileName, commitMetaData));
+		ProcessMetricTracker pmTracker = db.find(fileName, ProcessMetricTracker.class);
+		boolean exists = pmTracker != null;
+		if(!exists)
+			pmTracker = new ProcessMetricTracker(fileName, commitMetaData);
+
 		pmTracker.reportCommit(commitMetaData.getCommitId(), commitMetaData.getCommitMessage(), authorName, linesAdded, linesDeleted);
 
-		database.put(fileName, pmTracker);
+		if(exists)
+			db.update(pmTracker);
+		else
+			db.persist(pmTracker);
 	}
 
 	//Reset the tracker with latest refactoring and its commit meta data
 	//the commitCounter will be zero again
 	public void reportRefactoring(String fileName, CommitMetaData commitMetaData) {
-		ProcessMetricTracker pmTracker = database.getOrDefault(fileName, new ProcessMetricTracker(fileName, commitMetaData));
+		ProcessMetricTracker pmTracker = db.find(fileName, ProcessMetricTracker.class);
+		boolean exists = pmTracker != null;
+		if(!exists)
+			pmTracker = new ProcessMetricTracker(fileName, commitMetaData);
 		pmTracker.resetCounter(commitMetaData);
 
-		database.put(fileName, pmTracker);
+		if(exists)
+			db.update(pmTracker);
+		else
+			db.persist(pmTracker);
 	}
 
+	//TODO: adapt this to the new SQL database
+	@Deprecated
 	public String toString(){
 		return "PMDatabase{" +
-				"database=" + database.toString() + ",\n" +
+				"database=" + db.mapToString() + ",\n" +
 				"commitThreshold=" + commitThresholds.toString() +
 				"}";
 	}
